// src/app/models/ticket.ts
export class Ticket {
    "id": "string";
    "eventoId": number;
    "fecha": "string";
    "hora": number;
    "duracion": number;
    "precio": number;
    "silla": number
}

export class TicketSinID {
    "eventoId": number;
    "fecha": "string";
    "hora": number;
    "duracion": number;
    "precio": number;
    "silla": number
}